Sweetalert2

+++ DevBook/web/assets/js/publicacoes.js
@@ -19,7 +19,7 @@ function criarPublicacao(evento) {
   }).done(function() {
     window.location = "/home";
   }).fail(function() {
-    alert("Erro ao criar a publicação");
+    Swal.fire("Ops...", "Erro ao criar a publicação!", "error");
   });
 }
 
@@ -36,7 +36,6 @@ function curtirPublicacao(evento) {
       url: `/publicacoes/${publicacaoId}/curtir`,
       method: "POST"
   }).done(function() {
-    // alert("Publicação curtida");
     const contadorDeCurtidas = elementoClicado.next('span');
     const quantidadeDeCurtidas = parseInt(contadorDeCurtidas.text());
 
@@ -47,7 +46,7 @@ function curtirPublicacao(evento) {
     elementoClicado.removeClass('curtir-publicacao');
 
   }).fail(function() {
-    alert("Erro ao curtir publicação");
+    Swal.fire("Ops...", "Erro ao curtir publicação!", "error");
   }).always(function() {
     elementoClicado.prop('disabled', false);
   });
@@ -65,7 +64,6 @@ function descurtirPublicacao(evento) {
       url: `/publicacoes/${publicacaoId}/descurtir`,
       method: "POST"
   }).done(function() {
-    // alert("Publicação curtida");
     const contadorDeCurtidas = elementoClicado.next('span');
     const quantidadeDeCurtidas = parseInt(contadorDeCurtidas.text());
 
@@ -76,7 +74,7 @@ function descurtirPublicacao(evento) {
     elementoClicado.addClass('curtir-publicacao');
 
   }).fail(function() {
-    alert("Erro ao curtir publicação");
+    Swal.fire("Ops...", "Erro ao descurtir publicação!", "error");
   }).always(function() {
     elementoClicado.prop('disabled', false);
   });
@@ -96,9 +94,12 @@ function salvarPublicacao() {
       conteudo: $('#conteudo').val()
     }
   }).done(function() {
-    alert("Publicação atualizada");
+    Swal.fire(
+      "Sucesso!", "Publicação atualizada!", "success").then(function() {
+        window.location = "/home";
+      });
   }).fail(function() {
-    alert("Erro ao salvar a publicação");
+    Swal.fire("Ops...", "Erro ao salvar a publicação!", "error");
   }).always(function() {
     $('#salvar-publicacao').prop('disabled', false);
   });
@@ -107,22 +108,32 @@ function salvarPublicacao() {
 function deletarPublicacao(evento) {
   evento.preventDefault();
 
-  const elementoClicado = $(evento.target);
-  const publicacao = elementoClicado.closest('div');
-  const publicacaoId = publicacao.data('publicacao-id');
+  Swal.fire({
+    title: "Atenção!",
+    text: "Ação irreversível, deseja realmente excluir essa publicação?",
+    showCancelButton: true,
+    cancelButtomText: "Cancelar",
+    icon: "warning"
+  }).then(function(confirmacao) {
+    if (!confirmacao.value) return;
 
-  elementoClicado.prop('disabled', true);
-
-  $.ajax({
-    url: `/publicacoes/${publicacaoId}`,
-    method: "DELETE"
-  }).done(function() {
-    publicacao.fadeOut("slow", function() {
-      $(this).remove();
-    });
-  }).fail(function() {
-    alert("Erro ao excluir a publicação")
-  }).always(function() {
-    elementoClicado.prop('disabled', false);   
+    const elementoClicado = $(evento.target);
+    const publicacao = elementoClicado.closest('div');
+    const publicacaoId = publicacao.data('publicacao-id');
+  
+    elementoClicado.prop('disabled', true);
+  
+    $.ajax({
+      url: `/publicacoes/${publicacaoId}`,
+      method: "DELETE"
+    }).done(function() {
+      publicacao.fadeOut("slow", function() {
+        $(this).remove();
+      });
+    }).fail(function() {
+      Swal.fire("Ops...", "Erro ao excluir a publicação!", "error");
+    }).always(function() {
+      elementoClicado.prop('disabled', false);   
+    })
   })
 }
